backend/pyproject.toml
[project]
name = "data-extraction-tool"
version = "0.1.0"
description = "Scientific article data extraction with GRADE framework assessment"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "anthropic>=0.40.0",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "docxtpl>=0.18.0",
    "python-multipart>=0.0.12",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "pgvector>=0.3.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
